requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
PyMySQL==1.1.0
python-dotenv==1.0.0
email-validator==2.0.0
requests==2.31.0
cryptography==41.0.4
Werkzeug==2.3.7
types-requests==2.32.4.20250611
